Download a changeset in OsmChange format

Returns the OsmChange document describing all changes associated with the changeset.

Usage

osm_download_changeset(changeset_id, format = c("R", "xml"))

Arguments

changeset_id

The id of the changeset represented by a numeric or a character value for which the OsmChange is requested.

format

Format of the output. Can be R (default) or xml.

Details

The result of calling this may change as long as the changeset is open.

The elements in the OsmChange are sorted by timestamp and version number.

There is osm_read_changeset() to get only information about the changeset itself.

See also

Examples

if (FALSE) {
chaset <- osm_download_changeset(changeset_id = 137003062)
chaset
}