RSS Feed of notes in a bbox

RSS Feed of notes in a bbox

Usage

osm_feed_notes(bbox)

Arguments

bbox

Coordinates for the area to retrieve the notes from (left,bottom,right,top). Floating point numbers in degrees, expressing a valid bounding box.

See also

Examples

feed_notes <- osm_feed_notes(bbox = c(0.8205414, 40.6686604, 0.8857727, 40.7493377))
## bbox as a character value also works. Equivalent call:
# feed_notes <- osm_feed_notes(bbox = "0.8205414,40.6686604,0.8857727,40.7493377")
feed_notes
#> {xml_document}
#> <rss version="2.0" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:geo="http://www.w3.org/2003/01/geo/wgs84_pos#" xmlns:georss="http://www.georss.org/georss">
#> [1] <channel>\n  <title>OpenStreetMap Notes</title>\n  <description>A list of ...