Get the history of an object

Retrieves all old versions of an object from OSM, sorted by version number from oldest to newest.

Usage

osm_history_object(
  osm_type = c("node", "way", "relation"),
  osm_id,
  format = c("R", "xml", "json"),
  tags_in_columns = FALSE
)

Arguments

osm_type

Object type ("node", "way" or "relation").

osm_id

Object id represented by a numeric or a character value.

format

Format of the output. Can be "R" (default), "xml", or "json".

tags_in_columns

If FALSE (default), the tags of the objects are saved in a single list column tags containing a data.frame for each OSM object with the keys and values. If TRUE, add a column for each key. Ignored if format != "R".

Value

If format = "R", returns a data frame with a version of the OSM object per row. If format = "xml", returns a xml2::xml_document following the OSM_XML format. If format = "json", returns a list with a json structure following the OSM_JSON format.

See also

Examples

node <- osm_history_object(osm_type = "node", osm_id = 35308286)
node
#>    type       id visible version changeset           timestamp     user
#> 1  node 35308286    TRUE       1    292412 2007-09-02 14:55:33  Skywave
#> 2  node 35308286    TRUE       2     39997 2007-12-02 19:28:54  Skywave
#> 3  node 35308286    TRUE       3     39997 2007-12-02 19:29:01  Skywave
#> 4  node 35308286    TRUE       4    131335 2007-12-06 21:08:56  Skywave
#> 5  node 35308286    TRUE       5    544716 2008-01-05 17:55:28  Skywave
#> 6  node 35308286    TRUE       6    433656 2008-08-28 00:14:47  Skywave
#> 7  node 35308286    TRUE       7     61302 2008-10-04 20:55:22  Skywave
#> 8  node 35308286    TRUE       8    731089 2009-03-03 20:23:52   Eric S
#> 9  node 35308286    TRUE       9   4188890 2010-03-21 10:17:52 petrovsk
#> 10 node 35308286    TRUE      10   5978159 2010-10-07 13:28:57  Skywave
#> 11 node 35308286    TRUE      11   5980618 2010-10-07 19:16:33  Skywave
#> 12 node 35308286    TRUE      12   9467521 2011-10-04 08:48:09  WernerP
#> 13 node 35308286    TRUE      13  13343795 2012-10-03 07:49:42      FMB
#> 14 node 35308286    TRUE      14  18407951 2013-10-17 18:21:18    janes
#> 15 node 35308286    TRUE      15  63949815 2018-10-28 13:07:37   ManelG
#> 16 node 35308286    TRUE      16 140341359 2023-08-24 20:19:21 jmaspons
#> 17 node 35308286    TRUE      17 140341361 2023-08-24 20:19:22 jmaspons
#>         uid        lat       lon members
#> 1     10927 42.5178841 2.4567366    NULL
#> 2     10927 42.5116144 2.4580973    NULL
#> 3     10927 42.5184471 2.4580973    NULL
#> 4     10927 42.5184471 2.4580973    NULL
#> 5     10927 42.5184471 2.4580973    NULL
#> 6     10927 42.5184481 2.4580929    NULL
#> 7     10927 42.5187840 2.4567709    NULL
#> 8     45284 42.5187840 2.4567709    NULL
#> 9     90394 42.5189368 2.4566062    NULL
#> 10    10927 42.5189176 2.4565525    NULL
#> 11    10927 42.5189047 2.4565596    NULL
#> 12   315015 42.5189047 2.4565596    NULL
#> 13   316958 42.5189047 2.4565596    NULL
#> 14   428521 42.5189047 2.4565596    NULL
#> 15  2491053 42.5189047 2.4565596    NULL
#> 16 11725140 42.5189047 2.4565596    NULL
#> 17 11725140 42.5189047 2.4565596    NULL
#>                                                                                   tags
#> 1  5 tags: altitude=2784,66 | created_by=Potlatch alpha | ele=2784,66 | name=Pic du...
#> 2  5 tags: altitude=2784,66 | created_by=Potlatch 0.5d | ele=2784,66 | name=Pic du ...
#> 3  5 tags: altitude=2784,66 | created_by=Potlatch 0.5d | ele=2784,66 | name=Pic du ...
#> 4  4 tags: created_by=Potlatch 0.5d | ele=2784,66 | name=Pic du Canigou | natural=p...
#> 5  5 tags: created_by=Potlatch 0.5d | ele=2784,66 | name=Pic du Canigou | name:ca=P...
#> 6  6 tags: created_by=Potlatch 0.10b | ele=2784,66 | elevation=2784,66 meters | nam...
#> 7  6 tags: created_by=Potlatch 0.10c | ele=2784,66 | elevation=2784,66 meters | nam...
#> 8  4 tags: ele=2784,66 | name=Pic du Canigou | name:ca=Puig del Canigó | natural=pe...
#> 9  4 tags: ele=2784,66 | name=Pic du Canigou | name:ca=Puig del Canigó | natural=pe...
#> 10 4 tags: ele=2784,66 | name=Pic du Canigou | name:ca=Puig del Canigó | natural=pe...
#> 11 4 tags: ele=2784,66 | name=Pic du Canigou | name:ca=Puig del Canigó | natural=pe...
#> 12 4 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Puig del Canigó | natural=pe...
#> 13 4 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pica del Canigó | natural=pe...
#> 14 4 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...
#> 15 5 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...
#> 16 6 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...
#> 17 7 tags: ele=2784.66 | name=Pic du Canigou | name:ca=Pic del Canigó | natural=pea...

way <- osm_history_object(osm_type = "way", osm_id = 13073736L)
way
#>    type       id visible version changeset           timestamp
#> 1   way 13073736    TRUE       1    545236 2007-11-17 22:30:37
#> 2   way 13073736    TRUE       2    270381 2007-12-12 11:13:02
#> 3   way 13073736    TRUE       3   9724157 2011-11-02 17:55:59
#> 4   way 13073736    TRUE       4  12745372 2012-08-15 23:09:48
#> 5   way 13073736    TRUE       5  16749437 2013-06-29 10:06:43
#> 6   way 13073736    TRUE       6  24765034 2014-08-15 11:27:50
#> 7   way 13073736    TRUE       7  35855685 2015-12-09 22:11:41
#> 8   way 13073736    TRUE       8  44486884 2016-12-18 10:12:22
#> 9   way 13073736    TRUE       9  49340277 2017-06-07 15:51:46
#> 10  way 13073736    TRUE      10  55126121 2018-01-03 11:23:29
#> 11  way 13073736    TRUE      11  63379426 2018-10-10 12:39:54
#> 12  way 13073736    TRUE      12  74075591 2019-09-04 09:57:46
#> 13  way 13073736    TRUE      13  74385011 2019-09-12 09:40:46
#> 14  way 13073736    TRUE      14  75680406 2019-10-14 17:15:24
#> 15  way 13073736    TRUE      15  76002470 2019-10-21 14:21:08
#> 16  way 13073736    TRUE      16  76811325 2019-11-08 13:10:11
#> 17  way 13073736    TRUE      17  77447443 2019-11-22 22:35:38
#> 18  way 13073736    TRUE      18  86534163 2020-06-11 21:02:34
#> 19  way 13073736    TRUE      19  92154431 2020-10-08 07:42:45
#> 20  way 13073736    TRUE      20  92482152 2020-10-14 15:49:32
#> 21  way 13073736    TRUE      21 114656477 2021-12-07 11:27:19
#> 22  way 13073736    TRUE      22 140148279 2023-08-20 19:16:50
#> 23  way 13073736    TRUE      23 140150766 2023-08-20 20:29:55
#> 24  way 13073736    TRUE      24 140156836 2023-08-21 00:40:48
#> 25  way 13073736    TRUE      25 140161821 2023-08-21 05:35:09
#> 26  way 13073736    TRUE      26 140197966 2023-08-21 19:33:10
#> 27  way 13073736    TRUE      27 140205821 2023-08-22 01:47:19
#> 28  way 13073736    TRUE      28 158012509 2024-10-17 14:56:13
#> 29  way 13073736    TRUE      29 158012506 2024-10-17 14:56:13
#>                            user      uid  lat  lon
#> 1                   ivansanchez     5265 <NA> <NA>
#> 2                          jgui     8016 <NA> <NA>
#> 3                         morsi   139580 <NA> <NA>
#> 4                       rubensd   472193 <NA> <NA>
#> 5                    pere prlpz  1513644 <NA> <NA>
#> 6                        qientx  2258523 <NA> <NA>
#> 7                      gorogoro  2937174 <NA> <NA>
#> 8                 Леонид Библис  4979200 <NA> <NA>
#> 9                 Marion_Moseby  5379200 <NA> <NA>
#> 10                     av223119   190711 <NA> <NA>
#> 11                Baxe Nafarroa  6672338 <NA> <NA>
#> 12               homocomputeris  3777620 <NA> <NA>
#> 13                      rubensd   472193 <NA> <NA>
#> 14                CapitanColoto  8940402 <NA> <NA>
#> 15                CapitanColoto  8940402 <NA> <NA>
#> 16           editemapes_imports 10063912 <NA> <NA>
#> 17                Georg_M_Weber 10241965 <NA> <NA>
#> 18                    MikelCalo 11126976 <NA> <NA>
#> 19                 Emilio Gomez     2904 <NA> <NA>
#> 20                     jmaspons 11725140 <NA> <NA>
#> 21                     jmaspons 11725140 <NA> <NA>
#> 22                     CorfuF08 20047280 <NA> <NA>
#> 23                       Taya_S 15944521 <NA> <NA>
#> 24 OverRtliyBoomrYxtracalicular 20047746 <NA> <NA>
#> 25                    NorthCrab 15215305 <NA> <NA>
#> 26     UnfNYllyHyperdposityL736 20048566 <NA> <NA>
#> 27           SomeoneElse_Revert  1778799 <NA> <NA>
#> 28               homocomputeris  3777620 <NA> <NA>
#> 29               homocomputeris  3777620 <NA> <NA>
#>                                                            members
#> 1  17 nodes: 120303676, 120303671, 120303597, 120303605, 120303...
#> 2  17 nodes: 120303676, 120303671, 120303597, 120303605, 120303...
#> 3  17 nodes: 120303676, 120303671, 120303597, 120303605, 120303...
#> 4  17 nodes: 120303676, 120303671, 120303597, 120303605, 120303...
#> 5  17 nodes: 120303676, 120303671, 120303597, 120303605, 120303...
#> 6  19 nodes: 120303676, 120303671, 120303597, 3017814863, 30178...
#> 7  19 nodes: 120303676, 120303671, 120303597, 3017814863, 30178...
#> 8  19 nodes: 120303676, 120303671, 120303597, 3017814863, 30178...
#> 9  20 nodes: 120303676, 120303671, 120303597, 3017814863, 30178...
#> 10 20 nodes: 120303676, 120303671, 120303597, 3017814863, 30178...
#> 11 20 nodes: 120303676, 120303671, 120303597, 3017814863, 30178...
#> 12 34 nodes: 120303676, 120303671, 120303597, 4902243082, 12030...
#> 13 34 nodes: 120303676, 120303671, 120303597, 6790061967, 12030...
#> 14 34 nodes: 120303676, 120303671, 120303597, 6790061967, 12030...
#> 15 34 nodes: 120303676, 120303671, 120303597, 6790061967, 12030...
#> 16 59 nodes: 6771540804, 6771540805, 6771540806, 120303605, 695...
#> 17 60 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 18 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 19 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 20 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 21 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 22 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 23 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 24 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 25 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 26 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 27 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 28 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#> 29 61 nodes: 6771540804, 6771540805, 6771540806, 6957604952, 67...
#>                                                                                   tags
#> 1  4 tags: created_by=Potlatch alpha | historic=castle | name=Torres de Quart | tou...
#> 2                                   2 tags: building=tower | created_by=Potlatch alpha
#> 3                       3 tags: building=tower | historic=tower | name=Torres de Quart
#> 4  4 tags: building=tower | historic=tower | name=Torres de Quart | tourism=attract...
#> 5  5 tags: building=tower | historic=tower | name=Torres de Quart | tourism=attract...
#> 6  5 tags: building=tower | historic=tower | name=Torres de Quart | tourism=attract...
#> 7  7 tags: building=tower | building:material=stone | height=34 | historic=tower | ...
#> 8  8 tags: building=tower | building:material=stone | height=34 | historic=tower | ...
#> 9  8 tags: building=tower | building:material=stone | height=34 | historic=tower | ...
#> 10 9 tags: building=tower | building:material=stone | height=34 | historic=tower | ...
#> 11 8 tags: building=tower | building:material=stone | height=34 | historic=city_gat...
#> 12 11 tags: alt_name:ca=Torres de Quart | building=tower | building:material=stone ...
#> 13 11 tags: alt_name:ca=Torres de Quart | building=tower | building:material=stone ...
#> 14 10 tags: alt_name:ca=Torres de Quart | building=tower | building:material=stone ...
#> 15 8 tags: alt_name:ca=Torres de Quart | building=tower | building:material=stone |...
#> 16 9 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | build...
#> 17 9 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | build...
#> 18 9 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | build...
#> 19 10 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 20 11 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 21 12 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 22 11 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 23 12 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 24 11 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 25 12 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 26 11 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 27 12 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 28 13 tags: alt_name:ca=Torres de Quart | building=tower | building:levels=4 | buil...
#> 29 14 tags: alt_name:ca=Torres de Quart | building=tower | building:colour=beige | ...

rel <- osm_history_object(osm_type = "relation", osm_id = "40581")
rel
#>        type    id visible version changeset           timestamp
#> 1  relation 40581    TRUE       1    568221 2008-10-25 05:35:48
#> 2  relation 40581    TRUE       2  11714681 2012-05-27 12:05:32
#> 3  relation 40581    TRUE       3  11746232 2012-05-30 08:57:05
#> 4  relation 40581    TRUE       4  11837646 2012-06-08 18:01:23
#> 5  relation 40581    TRUE       5  11839138 2012-06-08 20:25:11
#> 6  relation 40581    TRUE       6  11838130 2012-06-08 20:51:42
#> 7  relation 40581    TRUE       7  11980761 2012-06-22 16:39:50
#> 8  relation 40581    TRUE       8  18369633 2013-10-15 14:16:51
#> 9  relation 40581    TRUE       9  18654817 2013-11-01 17:02:49
#> 10 relation 40581    TRUE      10  19864829 2014-01-07 15:19:27
#> 11 relation 40581    TRUE      11  19893801 2014-01-09 03:42:07
#> 12 relation 40581    TRUE      12  19904315 2014-01-09 17:49:09
#> 13 relation 40581    TRUE      13  19909641 2014-01-09 23:03:35
#> 14 relation 40581    TRUE      14  20324867 2014-02-01 21:58:59
#> 15 relation 40581    TRUE      15  20783674 2014-02-26 03:52:17
#> 16 relation 40581    TRUE      16  20783699 2014-02-26 03:56:58
#> 17 relation 40581    TRUE      17  25024005 2014-08-26 10:32:56
#> 18 relation 40581    TRUE      18  31684773 2015-06-03 07:36:51
#> 19 relation 40581    TRUE      19  40885359 2016-07-20 01:19:33
#> 20 relation 40581    TRUE      20  41147692 2016-07-31 15:16:38
#> 21 relation 40581    TRUE      21  41156298 2016-08-01 00:24:38
#> 22 relation 40581    TRUE      22  41162305 2016-08-01 09:03:50
#> 23 relation 40581    TRUE      23  41206301 2016-08-03 03:36:05
#> 24 relation 40581    TRUE      24  41212009 2016-08-03 11:16:49
#> 25 relation 40581    TRUE      25  41214566 2016-08-03 13:21:37
#> 26 relation 40581    TRUE      26  41214925 2016-08-03 13:40:57
#> 27 relation 40581    TRUE      27  41215848 2016-08-03 14:26:01
#> 28 relation 40581    TRUE      28  41216353 2016-08-03 14:50:48
#> 29 relation 40581    TRUE      29  41217947 2016-08-03 16:12:29
#> 30 relation 40581    TRUE      30  41221606 2016-08-03 18:54:32
#> 31 relation 40581    TRUE      31  41223121 2016-08-03 19:58:40
#> 32 relation 40581    TRUE      32  41223747 2016-08-03 20:26:15
#> 33 relation 40581    TRUE      33  41852822 2016-09-01 16:28:22
#> 34 relation 40581    TRUE      34  41869838 2016-09-02 14:05:24
#> 35 relation 40581    TRUE      35  41889305 2016-09-03 13:57:10
#> 36 relation 40581    TRUE      36  41891270 2016-09-03 15:50:28
#> 37 relation 40581    TRUE      37  43712169 2016-11-16 21:48:53
#> 38 relation 40581    TRUE      38  43742170 2016-11-17 20:03:24
#> 39 relation 40581    TRUE      39  43750924 2016-11-18 01:09:29
#> 40 relation 40581    TRUE      40  43854877 2016-11-21 19:18:40
#> 41 relation 40581    TRUE      41  43856068 2016-11-21 20:25:39
#> 42 relation 40581    TRUE      42  44927268 2017-01-05 15:17:57
#> 43 relation 40581    TRUE      43  51567307 2017-08-30 08:28:01
#> 44 relation 40581    TRUE      44  51609472 2017-08-31 10:50:49
#> 45 relation 40581    TRUE      45  66542092 2019-01-22 15:54:20
#> 46 relation 40581    TRUE      46  66543985 2019-01-22 17:01:40
#> 47 relation 40581    TRUE      47 107730109 2021-07-10 00:32:19
#> 48 relation 40581    TRUE      48 111244602 2021-09-15 13:05:46
#> 49 relation 40581    TRUE      49 115229896 2021-12-22 00:20:00
#> 50 relation 40581    TRUE      50 115501973 2021-12-29 06:50:31
#> 51 relation 40581    TRUE      51 118611187 2022-03-17 21:13:14
#> 52 relation 40581    TRUE      52 124236416 2022-07-29 15:11:43
#> 53 relation 40581    TRUE      53 135982205 2023-05-11 15:10:18
#> 54 relation 40581    TRUE      54 139761106 2023-08-11 18:02:05
#> 55 relation 40581    TRUE      55 143271690 2023-10-28 16:39:31
#> 56 relation 40581    TRUE      56 150504881 2024-04-25 16:28:13
#> 57 relation 40581    TRUE      57 150508547 2024-04-25 18:02:27
#> 58 relation 40581    TRUE      58 150515740 2024-04-25 22:10:14
#> 59 relation 40581    TRUE      59 150693583 2024-04-30 07:43:50
#> 60 relation 40581    TRUE      60 150702242 2024-04-30 11:14:53
#> 61 relation 40581    TRUE      61 150722050 2024-04-30 18:29:37
#> 62 relation 40581    TRUE      62 151100114 2024-05-09 13:23:23
#> 63 relation 40581    TRUE      63 153550023 2024-07-04 14:07:58
#> 64 relation 40581    TRUE      64 154460676 2024-07-26 20:34:28
#> 65 relation 40581    TRUE      65 154709112 2024-08-01 19:46:31
#> 66 relation 40581    TRUE      66 155092910 2024-08-11 09:34:56
#> 67 relation 40581    TRUE      67 155115281 2024-08-11 17:51:57
#> 68 relation 40581    TRUE      68 155968368 2024-08-30 11:14:10
#> 69 relation 40581    TRUE      69 156149027 2024-09-03 15:56:39
#> 70 relation 40581    TRUE      70 158189552 2024-10-21 22:17:57
#>                  user      uid  lat  lon
#> 1              simone      137 <NA> <NA>
#> 2               lerks   188477 <NA> <NA>
#> 3               lerks   188477 <NA> <NA>
#> 4             sabas88   454589 <NA> <NA>
#> 5               lerks   188477 <NA> <NA>
#> 6               lerks   188477 <NA> <NA>
#> 7             sabas88   454589 <NA> <NA>
#> 8               l2212  1775455 <NA> <NA>
#> 9               l2212  1775455 <NA> <NA>
#> 10           Paoletto   170997 <NA> <NA>
#> 11              l2212  1775455 <NA> <NA>
#> 12           Paoletto   170997 <NA> <NA>
#> 13              l2212  1775455 <NA> <NA>
#> 14       mcheckimport   893327 <NA> <NA>
#> 15              l2212  1775455 <NA> <NA>
#> 16              l2212  1775455 <NA> <NA>
#> 17           jmontane    28773 <NA> <NA>
#> 18             simone      137 <NA> <NA>
#> 19              fayor   743479 <NA> <NA>
#> 20              l2212  1775455 <NA> <NA>
#> 21              fayor   743479 <NA> <NA>
#> 22              l2212  1775455 <NA> <NA>
#> 23              fayor   743479 <NA> <NA>
#> 24              l2212  1775455 <NA> <NA>
#> 25              fayor   743479 <NA> <NA>
#> 26              l2212  1775455 <NA> <NA>
#> 27              fayor   743479 <NA> <NA>
#> 28              l2212  1775455 <NA> <NA>
#> 29              fayor   743479 <NA> <NA>
#> 30              l2212  1775455 <NA> <NA>
#> 31              fayor   743479 <NA> <NA>
#> 32              l2212  1775455 <NA> <NA>
#> 33              fayor   743479 <NA> <NA>
#> 34              l2212  1775455 <NA> <NA>
#> 35              fayor   743479 <NA> <NA>
#> 36 SomeoneElse_Revert  1778799 <NA> <NA>
#> 37           GBicozzo  4853063 <NA> <NA>
#> 38 SomeoneElse_Revert  1778799 <NA> <NA>
#> 39            nyuriks   339581 <NA> <NA>
#> 40           GBicozzo  4853063 <NA> <NA>
#> 41 SomeoneElse_Revert  1778799 <NA> <NA>
#> 42             dan980  1240942 <NA> <NA>
#> 43           osm_fede  5631066 <NA> <NA>
#> 44        user_177389   177389 <NA> <NA>
#> 45            s141739    25132 <NA> <NA>
#> 46        user_177389   177389 <NA> <NA>
#> 47              Olyon  1443767 <NA> <NA>
#> 48           nilogian   673435 <NA> <NA>
#> 49             Yamílo 11332999 <NA> <NA>
#> 50          user_5359     5359 <NA> <NA>
#> 51     Gianfranco2014  1928626 <NA> <NA>
#> 52           ftrebien   401472 <NA> <NA>
#> 53              l2212  1775455 <NA> <NA>
#> 54           jmaspons 11725140 <NA> <NA>
#> 55         b-jazz-bot  9451067 <NA> <NA>
#> 56        Siciliapars 21349271 <NA> <NA>
#> 57            cucunto 21350717 <NA> <NA>
#> 58        Siciliapars 21349271 <NA> <NA>
#> 59             froipe 21385987 <NA> <NA>
#> 60         Adriaticum 21387693 <NA> <NA>
#> 61             froipe 21385987 <NA> <NA>
#> 62        Siciliapars 21349271 <NA> <NA>
#> 63               aboc 21696131 <NA> <NA>
#> 64           jmaspons 11725140 <NA> <NA>
#> 65               aboc 21696131 <NA> <NA>
#> 66               aboc 21696131 <NA> <NA>
#> 67        user_177389   177389 <NA> <NA>
#> 68               aboc 21696131 <NA> <NA>
#> 69           jmaspons 11725140 <NA> <NA>
#> 70               aboc 21696131 <NA> <NA>
#>                                                            members
#> 1  7 members: way/27966652/, way/27966705/, way/27966846/, way/...
#> 2  7 members: way/27966652/outer, way/27966705/outer, way/27966...
#> 3  7 members: way/27966652/outer, way/27966705/outer, way/27966...
#> 4  14 members: way/27966652/outer, way/27966705/outer, way/2796...
#> 5  13 members: way/27966652/outer, way/27966846/outer, way/2796...
#> 6  13 members: way/27966846/outer, way/27966652/outer, way/1666...
#> 7  13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 8  13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 9  13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 10 13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 11 13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 12 13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 13 13 members: way/166623039/outer, way/165579897/outer, way/11...
#> 14 14 members: way/166623039/outer, way/165579897/outer, way/11...
#> 15 14 members: way/166623039/outer, way/165579897/outer, way/11...
#> 16 14 members: way/166623039/outer, way/165579897/outer, way/11...
#> 17 14 members: way/166623039/outer, way/165579897/outer, way/11...
#> 18 14 members: way/166623039/outer, way/165579897/outer, way/11...
#> 19 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 20 14 members: way/166623039/outer, way/165579897/outer, way/11...
#> 21 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 22 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 23 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 24 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 25 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 26 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 27 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 28 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 29 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 30 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 31 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 32 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 33 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 34 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 35 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 36 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 37 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 38 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 39 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 40 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 41 14 members: way/166623039/outer, way/168606149/outer, way/16...
#> 42 15 members: way/166623039/outer, way/168606149/outer, way/16...
#> 43 15 members: way/166623039/outer, way/168606149/outer, way/16...
#> 44 15 members: way/166623039/outer, way/168606149/outer, way/16...
#> 45 15 members: way/166623039/outer, way/168606149/outer, way/16...
#> 46 15 members: way/166623039/outer, way/168606149/outer, way/16...
#> 47 16 members: way/165579897/outer, way/963065356/outer, way/16...
#> 48 18 members: node/72994199/admin_centre, way/165579897/outer,...
#> 49 18 members: node/72994199/admin_centre, way/165579897/outer,...
#> 50 18 members: node/72994199/admin_centre, way/165579897/outer,...
#> 51 18 members: node/72994199/admin_centre, way/165579897/outer,...
#> 52 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 53 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 54 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 55 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 56 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 57 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 58 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 59 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 60 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 61 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 62 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 63 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 64 20 members: node/72994199/admin_centre, way/165579897/outer,...
#> 65 21 members: node/72994199/admin_centre, way/165579897/outer,...
#> 66 23 members: node/72994199/admin_centre, way/165579897/outer,...
#> 67 23 members: node/72994199/admin_centre, way/165579897/outer,...
#> 68 25 members: node/72994199/admin_centre, way/165579897/outer,...
#> 69 25 members: node/72994199/admin_centre, way/165579897/outer,...
#> 70 25 members: node/72994199/admin_centre, way/165579897/outer,...
#>                                                                                   tags
#> 1  6 tags: admin_level=8 | boundary=administrative | created_by=ShapeToOSM - Laser ...
#> 2  5 tags: admin_level=8 | boundary=administrative | name=Alghero | source=Based on...
#> 3  7 tags: admin_level=8 | boundary=administrative | name=Alghero | ref:catasto=A19...
#> 4  7 tags: admin_level=8 | boundary=administrative | name=Alghero | ref:catasto=A19...
#> 5  7 tags: admin_level=8 | boundary=administrative | name=Alghero | ref:catasto=A19...
#> 6  7 tags: admin_level=8 | boundary=administrative | name=Alghero | ref:catasto=A19...
#> 7  7 tags: admin_level=8 | boundary=administrative | name=Alghero | ref:catasto=A19...
#> 8  10 tags: admin_level=8 | boundary=administrative | name=L'Alguer-Alghero | name:...
#> 9  11 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 10 12 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 11 11 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 12 12 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 13 11 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 14 11 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 15 13 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 16 15 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 17 15 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 18 13 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 19 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 20 13 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 21 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 22 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 23 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 24 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 25 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 26 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 27 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 28 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 29 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 30 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 31 13 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 32 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 33 14 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 34 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 35 14 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 36 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 37 14 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 38 14 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 39 15 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 40 15 tags: admin_level=8 | boundary=administrative | name=Alghero | name:ca=L'Algu...
#> 41 15 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 42 15 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 43 13 tags: admin_level=8 | name=L'Alguer/Alghero | name:ca=L'Alguer | name:co=L'Al...
#> 44 15 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 45 15 tags: admin_level=8 | name=Alghero - L'Alguer | name:ca=L'Alguer | name:co=L'...
#> 46 15 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 47 15 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 48 15 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 49 24 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 50 21 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 51 22 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 52 22 tags: admin_level=8 | boundary=administrative | name=Alghero - L'Alguer | nam...
#> 53 22 tags: admin_level=8 | boundary=administrative | name=L'Alguer/Alghero | name:...
#> 54 22 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 55 22 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 56 22 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 57 22 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 58 22 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 59 22 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 60 22 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 61 22 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 62 22 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 63 23 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 64 24 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 65 24 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 66 24 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 67 24 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 68 24 tags: admin_level=8 | boundary=administrative | name=Alghero / l'Alguer | nam...
#> 69 24 tags: admin_level=8 | boundary=administrative | name=l'Alguer/Alghero | name:...
#> 70 26 tags: admin_level=8 | boundary=administrative | int_name=Alghero | name=L'Alg...