Relations of an object

Returns all (not deleted) relations in which the given object is used.

Usage

osm_relations_object(
  osm_type = c("node", "way", "relation"),
  osm_id,
  format = c("R", "xml", "json"),
  tags_in_columns = FALSE
)

Arguments

osm_type

Object type ("node", "way" or "relation").

osm_id

Object id represented by a numeric or a character value.

format

Format of the output. Can be "R" (default), "xml", or "json".

tags_in_columns

If FALSE (default), the tags of the objects are saved in a single list column tags containing a data.frame for each OSM object with the keys and values. If TRUE, add a column for each key. Ignored if format != "R".

Value

If format = "R", returns a data frame with one OSM object per row. If format = "xml", returns a xml2::xml_document following the OSM_XML format. If format = "json", returns a list with a json structure following the OSM_JSON format.

See also

Examples

node <- osm_relations_object(osm_type = "node", osm_id = 152364165)
node
#>       type       id visible version changeset           timestamp   user
#> 1 relation   347950    TRUE     106 152390011 2024-06-07 19:07:15 EliziR
#> 2 relation   349035    TRUE     167 152390043 2024-06-07 19:08:20 EliziR
#> 3 relation   349053    TRUE     800 153268403 2024-06-27 22:19:52   sbre
#> 4 relation  2417889    TRUE      59 152390011 2024-06-07 19:07:15 EliziR
#> 5 relation  2807397    TRUE     111 152390043 2024-06-07 19:08:20 EliziR
#> 6 relation 11746917    TRUE     223 152390043 2024-06-07 19:08:20 EliziR
#> 7 relation 11747082    TRUE     405 152390043 2024-06-07 19:08:20 EliziR
#>        uid  lat  lon
#> 1   605366 <NA> <NA>
#> 2   605366 <NA> <NA>
#> 3 19863282 <NA> <NA>
#> 4   605366 <NA> <NA>
#> 5   605366 <NA> <NA>
#> 6   605366 <NA> <NA>
#> 7   605366 <NA> <NA>
#>                                                           members
#> 1 97 members: node/152364165/admin_centre, way/96647329/outer,...
#> 2 515 members: node/152364165/admin_centre, way/45319680/outer...
#> 3 2229 members: node/152364165/admin_centre, node/602512999/la...
#> 4 113 members: node/152364165/admin_centre, way/185724656/oute...
#> 5 449 members: node/152364165/admin_centre, way/45327666/outer...
#> 6 2203 members: node/152364165/admin_centre, way/585618147/out...
#> 7 3497 members: node/8000963558/label, node/152364165/admin_ce...
#>                                                                                  tags
#> 1 33 tags: admin_level=8 | alt_name:gl=Barna | boundary=administrative | idee:name...
#> 2 31 tags: admin_level=6 | boundary=administrative | ine:provincia=08 | ISO3166-2=...
#> 3 67 tags: admin_level=4 | alt_name:el=Καταλονία | alt_name:eo=Katalunujo | alt_na...
#> 4 23 tags: admin_level=7 | boundary=administrative | idescat:comarca=13 | name=Bar...
#> 5 13 tags: boundary=political | idescat:àmbit=01 | name=Àmbit metropolità | name:a...
#> 6 8 tags: boundary=political | default_language=ca | name=Català com a llengua prò...
#> 7 44 tags: boundary=political | name=Països Catalans | name:an=Países Catalans | n...

way <- osm_relations_object(osm_type = "way", osm_id = 372011578)
way
#>       type      id visible version changeset           timestamp   user    uid
#> 1 relation 5524720    TRUE      38 153218834 2024-06-26 19:49:19 EliziR 605366
#>    lat  lon                                                         members
#> 1 <NA> <NA> 174 members: way/372006138/, way/372006137/, way/486335278/,...
#>                                                                                  tags
#> 1 9 tags: distance=62.65 | name=GR 179 Sender dels Maquis | network=nwn | operator...

rel <- osm_relations_object(osm_type = "relation", osm_id = 342792)
rel
#>       type       id visible version changeset           timestamp
#> 1 relation   349012    TRUE     108 148725656 2024-03-16 10:50:43
#> 2 relation 11739086    TRUE     109 148820542 2024-03-18 14:38:52
#>                user      uid  lat  lon
#> 1 Hugoren Martinako  3392826 <NA> <NA>
#> 2          Langenb1 20134577 <NA> <NA>
#>                                                           members
#> 1 392 members: node/21323935/admin_centre, way/86411973/outer,...
#> 2 1032 members: node/34105607/admin_centre, node/8000963555/la...
#>                                                                                  tags
#> 1 36 tags: admin_level=6 | alt_name:gl=Alacante | border_type=province | boundary=...
#> 2 54 tags: alt_name=País Valencià | alt_name:ca=País Valencià | boundary=political...