Search for notes

Returns the existing notes matching either the initial note text or any of the comments.

Usage

osm_search_notes(
  q,
  user,
  from,
  to,
  closed = 7,
  sort = c("updated_at", "created_at"),
  order = c("newest", "oldest"),
  limit = 100,
  format = c("data.frame", "xml", "rss", "json", "gpx")
)

Arguments

q

Specifies the search query.

user

Find notes by the user with the given user id (numeric) or display name (character).

from

Specifies the beginning of a date range to search in for a note. A valid ISO 8601 date.

to

Specifies the end of a date range to search in for a note. A valid ISO 8601 date. The date of today is the default.

closed

Specifies the number of days a note needs to be closed to no longer be returned. A value of 0 means only open notes are returned. A value of -1 means all notes are returned. 7 is the default.

sort

Specifies the value which should be used to sort the notes. It is either possible to sort them by their creation date (created_at) or the date of the last update (updated_at, the default).

order

Specifies the order of the returned notes. It is possible to order them in ascending (oldest) or descending (newest, the default) order.

limit

Specifies the number of entries returned at max. A value of between 1 and 10000 is valid. 100 is the default.

format

Format of the the returned list of notes. Can be data.frame (default), xml, rss, json or gpx.

Details

The notes will be ordered by the date of their last change, the most recent one will be first. If no query was specified, the latest notes are returned.

See also

Examples

notes <- osm_search_notes(q = "POI", from = "2017-10-01", to = "2017-10-27T15:27A", limit = 10)
notes
#> [1] id           url          comment_url  close_url    date_created
#> [6] status       comments    
#> <0 rows> (or 0-length row.names)

my_notes <- osm_search_notes(user = "jmaspons", closed = -1, format = "json")
my_notes
#> $type
#> [1] "FeatureCollection"
#> 
#> $features
#> list()
#>