Search for notes

Returns notes that match the specified query. If no query is provided, the most recently updated notes are returned.

Usage

osm_search_notes(
  q,
  user,
  bbox,
  from,
  to,
  closed = 7,
  sort = c("updated_at", "created_at"),
  order = c("newest", "oldest"),
  limit = 100,
  format = c("R", "xml", "rss", "json", "gpx")
)

Arguments

q

Text search query, matching either note text or comments.

user

Search for notes which the given user interacted with. The value can be the user id (numeric) or the display name (character).

bbox

Search area expressed as a string of coordinates of a valid bounding box (left,bottom,right,top) in decimal degrees. Area must be at most 25 square degrees (see osm_capabilities()$note_area and this line in settings for the current value).

from

Beginning date range for created_at or updated_at (specified by sort). Preferably in ISO 8601 date format.

to

End date range for created_at or updated_at (specified by sort). Preferably in ISO 8601 date format. Only works when from is supplied.

closed

Specifies the number of days a note needs to be closed to no longer be returned. A value of 0 means only open notes are returned. A value of -1 means all notes are returned. 7 is the default.

sort

Sort results by creation (created_at) or update date (updated_at, the default).

order

Sorting order. oldest is ascending order, newest is descending order (the default).

limit

Maximum number of results between 1 and 10000 (may change, see osm_capabilities()$api$notes for the current value). Default to 100.

format

Format of the the returned list of notes. Can be R (default), xml, rss, json or gpx.

Value

If format = "R", returns a data frame with one map note per row. If format = "json", returns a list with the json structure. For format in xml, rss, and gpx, a xml2::xml_document with the corresponding format.

Details

The notes will be ordered by the date of their last change, the most recent one will be first.

See also

Examples

notes <- osm_search_notes(
  q = "POI", bbox = "0.1594133,40.5229822,3.3222508,42.8615226",
  from = "2017-10-01", to = "2018-10-27T15:27A", limit = 10
)
notes
#> [1] lon          lat          id           url          comment_url 
#> [6] close_url    date_created status       comments    
#> <0 rows> (or 0-length row.names)

my_notes <- osm_search_notes(user = "jmaspons", closed = -1, format = "json")
my_notes
#> $type
#> [1] "FeatureCollection"
#> 
#> $features
#> $features[[1]]
#> $features[[1]]$type
#> [1] "Feature"
#> 
#> $features[[1]]$geometry
#> $features[[1]]$geometry$type
#> [1] "Point"
#> 
#> $features[[1]]$geometry$coordinates
#> $features[[1]]$geometry$coordinates[[1]]
#> [1] 2.18233
#> 
#> $features[[1]]$geometry$coordinates[[2]]
#> [1] 41.38373
#> 
#> 
#> 
#> $features[[1]]$properties
#> $features[[1]]$properties$id
#> [1] 53727
#> 
#> $features[[1]]$properties$url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/53727.json"
#> 
#> $features[[1]]$properties$comment_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/53727/comment.json"
#> 
#> $features[[1]]$properties$close_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/53727/close.json"
#> 
#> $features[[1]]$properties$date_created
#> [1] "2023-12-30 09:53:14 UTC"
#> 
#> $features[[1]]$properties$status
#> [1] "open"
#> 
#> $features[[1]]$properties$comments
#> $features[[1]]$properties$comments[[1]]
#> $features[[1]]$properties$comments[[1]]$date
#> [1] "2023-12-30 09:53:14 UTC"
#> 
#> $features[[1]]$properties$comments[[1]]$uid
#> [1] 12141
#> 
#> $features[[1]]$properties$comments[[1]]$user
#> [1] "jmaspons"
#> 
#> $features[[1]]$properties$comments[[1]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[1]]$properties$comments[[1]]$action
#> [1] "opened"
#> 
#> $features[[1]]$properties$comments[[1]]$text
#> [1] "Testing osmapiR"
#> 
#> $features[[1]]$properties$comments[[1]]$html
#> [1] "<p>Testing osmapiR</p>"
#> 
#> 
#> $features[[1]]$properties$comments[[2]]
#> $features[[1]]$properties$comments[[2]]$date
#> [1] "2023-12-30 09:53:24 UTC"
#> 
#> $features[[1]]$properties$comments[[2]]$uid
#> [1] 12141
#> 
#> $features[[1]]$properties$comments[[2]]$user
#> [1] "jmaspons"
#> 
#> $features[[1]]$properties$comments[[2]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[1]]$properties$comments[[2]]$action
#> [1] "closed"
#> 
#> $features[[1]]$properties$comments[[2]]$text
#> [1] ""
#> 
#> $features[[1]]$properties$comments[[2]]$html
#> [1] "<p></p>"
#> 
#> 
#> $features[[1]]$properties$comments[[3]]
#> $features[[1]]$properties$comments[[3]]$date
#> [1] "2023-12-30 09:55:05 UTC"
#> 
#> $features[[1]]$properties$comments[[3]]$uid
#> [1] 12141
#> 
#> $features[[1]]$properties$comments[[3]]$user
#> [1] "jmaspons"
#> 
#> $features[[1]]$properties$comments[[3]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[1]]$properties$comments[[3]]$action
#> [1] "reopened"
#> 
#> $features[[1]]$properties$comments[[3]]$text
#> [1] ""
#> 
#> $features[[1]]$properties$comments[[3]]$html
#> [1] "<p></p>"
#> 
#> 
#> 
#> 
#> 
#> $features[[2]]
#> $features[[2]]$type
#> [1] "Feature"
#> 
#> $features[[2]]$geometry
#> $features[[2]]$geometry$type
#> [1] "Point"
#> 
#> $features[[2]]$geometry$coordinates
#> $features[[2]]$geometry$coordinates[[1]]
#> [1] 2.18233
#> 
#> $features[[2]]$geometry$coordinates[[2]]
#> [1] 41.38373
#> 
#> 
#> 
#> $features[[2]]$properties
#> $features[[2]]$properties$id
#> [1] 53726
#> 
#> $features[[2]]$properties$url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/53726.json"
#> 
#> $features[[2]]$properties$comment_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/53726/comment.json"
#> 
#> $features[[2]]$properties$close_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/53726/close.json"
#> 
#> $features[[2]]$properties$date_created
#> [1] "2023-12-30 09:29:45 UTC"
#> 
#> $features[[2]]$properties$status
#> [1] "open"
#> 
#> $features[[2]]$properties$comments
#> $features[[2]]$properties$comments[[1]]
#> $features[[2]]$properties$comments[[1]]$date
#> [1] "2023-12-30 09:29:45 UTC"
#> 
#> $features[[2]]$properties$comments[[1]]$action
#> [1] "opened"
#> 
#> $features[[2]]$properties$comments[[1]]$text
#> [1] "Testing osmapiR"
#> 
#> $features[[2]]$properties$comments[[1]]$html
#> [1] "<p>Testing osmapiR</p>"
#> 
#> 
#> $features[[2]]$properties$comments[[2]]
#> $features[[2]]$properties$comments[[2]]$date
#> [1] "2023-12-30 09:43:44 UTC"
#> 
#> $features[[2]]$properties$comments[[2]]$uid
#> [1] 12141
#> 
#> $features[[2]]$properties$comments[[2]]$user
#> [1] "jmaspons"
#> 
#> $features[[2]]$properties$comments[[2]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[2]]$properties$comments[[2]]$action
#> [1] "commented"
#> 
#> $features[[2]]$properties$comments[[2]]$text
#> [1] "A new comment to the note"
#> 
#> $features[[2]]$properties$comments[[2]]$html
#> [1] "<p>A new comment to the note</p>"
#> 
#> 
#> 
#> 
#> 
#> $features[[3]]
#> $features[[3]]$type
#> [1] "Feature"
#> 
#> $features[[3]]$geometry
#> $features[[3]]$geometry$type
#> [1] "Point"
#> 
#> $features[[3]]$geometry$coordinates
#> $features[[3]]$geometry$coordinates[[1]]
#> [1] 0.1702526
#> 
#> $features[[3]]$geometry$coordinates[[2]]
#> [1] 40.73274
#> 
#> 
#> 
#> $features[[3]]$properties
#> $features[[3]]$properties$id
#> [1] 42091
#> 
#> $features[[3]]$properties$url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/42091.json"
#> 
#> $features[[3]]$properties$comment_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/42091/comment.json"
#> 
#> $features[[3]]$properties$close_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/42091/close.json"
#> 
#> $features[[3]]$properties$date_created
#> [1] "2023-07-24 08:15:21 UTC"
#> 
#> $features[[3]]$properties$status
#> [1] "open"
#> 
#> $features[[3]]$properties$comments
#> $features[[3]]$properties$comments[[1]]
#> $features[[3]]$properties$comments[[1]]$date
#> [1] "2023-07-24 08:15:21 UTC"
#> 
#> $features[[3]]$properties$comments[[1]]$uid
#> [1] 12141
#> 
#> $features[[3]]$properties$comments[[1]]$user
#> [1] "jmaspons"
#> 
#> $features[[3]]$properties$comments[[1]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[3]]$properties$comments[[1]]$action
#> [1] "opened"
#> 
#> $features[[3]]$properties$comments[[1]]$text
#> [1] "There is survey point here."
#> 
#> $features[[3]]$properties$comments[[1]]$html
#> [1] "<p>There is survey point here.</p>"
#> 
#> 
#> $features[[3]]$properties$comments[[2]]
#> $features[[3]]$properties$comments[[2]]$date
#> [1] "2023-07-24 08:31:34 UTC"
#> 
#> $features[[3]]$properties$comments[[2]]$uid
#> [1] 12141
#> 
#> $features[[3]]$properties$comments[[2]]$user
#> [1] "jmaspons"
#> 
#> $features[[3]]$properties$comments[[2]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[3]]$properties$comments[[2]]$action
#> [1] "commented"
#> 
#> $features[[3]]$properties$comments[[2]]$text
#> [1] "Right, add it."
#> 
#> $features[[3]]$properties$comments[[2]]$html
#> [1] "<p>Right, add it.</p>"
#> 
#> 
#> $features[[3]]$properties$comments[[3]]
#> $features[[3]]$properties$comments[[3]]$date
#> [1] "2023-07-24 08:36:15 UTC"
#> 
#> $features[[3]]$properties$comments[[3]]$uid
#> [1] 12141
#> 
#> $features[[3]]$properties$comments[[3]]$user
#> [1] "jmaspons"
#> 
#> $features[[3]]$properties$comments[[3]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[3]]$properties$comments[[3]]$action
#> [1] "closed"
#> 
#> $features[[3]]$properties$comments[[3]]$text
#> [1] ""
#> 
#> $features[[3]]$properties$comments[[3]]$html
#> [1] "<p></p>"
#> 
#> 
#> $features[[3]]$properties$comments[[4]]
#> $features[[3]]$properties$comments[[4]]$date
#> [1] "2023-07-24 08:37:16 UTC"
#> 
#> $features[[3]]$properties$comments[[4]]$uid
#> [1] 12141
#> 
#> $features[[3]]$properties$comments[[4]]$user
#> [1] "jmaspons"
#> 
#> $features[[3]]$properties$comments[[4]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[3]]$properties$comments[[4]]$action
#> [1] "reopened"
#> 
#> $features[[3]]$properties$comments[[4]]$text
#> [1] ""
#> 
#> $features[[3]]$properties$comments[[4]]$html
#> [1] "<p></p>"
#> 
#> 
#> 
#> 
#> 
#> $features[[4]]
#> $features[[4]]$type
#> [1] "Feature"
#> 
#> $features[[4]]$geometry
#> $features[[4]]$geometry$type
#> [1] "Point"
#> 
#> $features[[4]]$geometry$coordinates
#> $features[[4]]$geometry$coordinates[[1]]
#> [1] 0.1702526
#> 
#> $features[[4]]$geometry$coordinates[[2]]
#> [1] 40.73274
#> 
#> 
#> 
#> $features[[4]]$properties
#> $features[[4]]$properties$id
#> [1] 42090
#> 
#> $features[[4]]$properties$url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/42090.json"
#> 
#> $features[[4]]$properties$comment_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/42090/comment.json"
#> 
#> $features[[4]]$properties$close_url
#> [1] "https://master.apis.dev.openstreetmap.org/api/0.6/notes/42090/close.json"
#> 
#> $features[[4]]$properties$date_created
#> [1] "2023-07-24 08:15:15 UTC"
#> 
#> $features[[4]]$properties$status
#> [1] "open"
#> 
#> $features[[4]]$properties$comments
#> $features[[4]]$properties$comments[[1]]
#> $features[[4]]$properties$comments[[1]]$date
#> [1] "2023-07-24 08:15:15 UTC"
#> 
#> $features[[4]]$properties$comments[[1]]$uid
#> [1] 12141
#> 
#> $features[[4]]$properties$comments[[1]]$user
#> [1] "jmaspons"
#> 
#> $features[[4]]$properties$comments[[1]]$user_url
#> [1] "https://master.apis.dev.openstreetmap.org/user/jmaspons"
#> 
#> $features[[4]]$properties$comments[[1]]$action
#> [1] "opened"
#> 
#> $features[[4]]$properties$comments[[1]]$text
#> [1] "There is survey point here."
#> 
#> $features[[4]]$properties$comments[[1]]$html
#> [1] "<p>There is survey point here.</p>"
#> 
#> 
#> 
#> 
#> 
#>